archived-duplicates/pyairtable-automation-services/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# File processing
python-multipart==0.0.6

# Optional file processing libraries
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
pdfplumber==0.10.3

# Task scheduling
croniter==2.0.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# OpenTelemetry dependencies
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-system-metrics==0.42b0
opentelemetry-distro==0.42b0
opentelemetry-propagator-b3==1.21.0
opentelemetry-propagator-jaeger==1.21.0
opentelemetry-semantic-conventions==0.42b0

# Structured logging for Loki integration
structlog==23.2.0